+++ src/embeds/index.js
@@ -1,5 +1,9 @@
 // @flow
 
+import { reset as resetLocalStorage } from 'lib/localStorage';
+
+resetLocalStorage();
+
 // $FlowFixMe
 const load = require('promise?global!./bootstrap'); // eslint-disable-line import/no-webpack-loader-syntax