+++ src/index.js
@@ -1,5 +1,9 @@
 // @flow
 
+import { reset as resetLocalStorage } from 'lib/localStorage';
+
+resetLocalStorage();
+
 import ReactDOM from 'react-dom';
 import Router from './Router';
 import Base from './Base';